requirements.txt
opencv-python
mediapipe
pyautogui
pyqt5
pytesseract
pyttsx3
pillow
pyperclip
